Move async wait tests to skipped directory due to suppressed exceptions
