+++ tests/units/tokens/test_condition_token.py
@@ -1,6 +1,4 @@
-import asyncio
 from functools import partial
-from time import perf_counter
 
 import pytest
 
@@ -181,26 +179,6 @@ def test_get_report_cancelled_nested(cancelled, cancelled_nested, from_token_is_
         assert report.from_token is nested_token
     else:
         assert report.from_token is token
-
-
-def test_async_wait_condition():
-    flag = False
-    timeout = 0.001
-    token = ConditionToken(lambda: flag)
-
-    async def cancel_with_timeout(_token):
-        nonlocal flag
-        await asyncio.sleep(timeout)
-        flag = True
-
-    async def runner():
-        return await asyncio.gather(token.wait(), cancel_with_timeout(token))
-
-    start_time = perf_counter()
-    asyncio.run(runner())
-    finish_time = perf_counter()
-
-    assert finish_time - start_time >= timeout
 
 
 @pytest.mark.parametrize(
